Add users and roles
Users are uniquely identifiable by their email address. This is
enforced with a unique index in their migration.

Roles are stored in a separate model that joins the user model.
Users may have zero or more roles. The roles are stored in an
Ecto.Enum that enforces the values as :member or :pal, but if
new roles are envisioned later on then this can be easily expanded.