Ensure that the initial balance is set on the changeset, not just the database table.
I'm making an assumption that when users sign up they will have a
certain amount of free initial balance to use for requesting visits.
I previously set that value as the default value in the database migration
but I failed to set it as the default in the changeset as well, which means
that if we called `Users.create_user()` then it would return a `User{}` with
the `balance` field set to `nil`.