Add transactions, and allow pals to perform visits for members.
This transfers the visit's minute value from the member to the
pal, subtracting 15% and also rounding down. We create a transaction
as a log of the event.

This is implemented using an Ecto.Multi and run through a single
repo transaction, so if anything fails then the entire transaction
should fail and not affect anyone's balances.